chore: refactor ArmorValues
