+++ src/types/Character/ArmorValues/FullArmorValues.ts
@@ -1,103 +1,6 @@
-// TODO Deprecate once gear is fully supported
-
-import Size, { sizeCombatMod } from './Size';
-import Stats from './Stats';
-
-interface ArmorValuesBuilder {
-    armor?: number;
-    shield?: number;
-    dodge?: number;
-    nat?: number;
-    defl?: number;
-    misc?: number;
-    miscP?: number;
-    miscE?: number;
-    temp?: number;
-    tempP?: number;
-    tempE?: number;
-}
-
-type FullArmorValuesBuilder = ArmorValuesBuilder & {
-    str?: number,
-    dex?: number,
-    bab?: number,
-    size?: number,
-};
-
-type FullFromBaseBuilder = {
-    base: ArmorValues,
-    stats: Stats,
-    bab: number,
-    size: Size,
-};
-
-class ArmorValues {
-    private _armor: number;
-    private _shield: number;
-    private _dodge: number;
-    private _nat: number;
-    private _defl: number;
-    private _misc: number;
-    private _miscP: number; // Misc bonus - Physical (not touch)
-    private _miscE: number; // Misc bonus - Evasion (not ff)
-    private _temp: number;
-    private _tempP: number; // Temp bonus - Physical (not touch)
-    private _tempE: number; // Temp bonus - Evasion (not ff)
-
-    constructor({
-        armor = 0,
-        shield = 0,
-        dodge = 0,
-        nat = 0,
-        defl = 0,
-        misc = 0,
-        miscP = 0,
-        miscE = 0,
-        temp = 0,
-        tempP = 0,
-        tempE = 0,
-    }: ArmorValuesBuilder) {
-        this._armor = armor;
-        this._shield = shield;
-        this._dodge = dodge;
-        this._nat = nat;
-        this._defl = defl;
-        this._misc = misc;
-        this._miscP = miscP;
-        this._miscE = miscE;
-        this._temp = temp;
-        this._tempP = tempP;
-        this._tempE = tempE;
-    }
-
-    get armor(): number { return this._armor; }
-
-    get shield(): number { return this._shield; }
-
-    get dodge(): number { return this._dodge; }
-
-    get nat(): number { return this._nat; }
-
-    get defl(): number { return this._defl; }
-
-    get misc(): number { return this._misc; }
-
-    get miscP(): number { return this._miscP; }
-
-    get miscE(): number { return this._miscE; }
-
-    get miscAll(): number { return this.misc + this.miscP + this.miscE; }
-
-    get temp(): number { return this._temp; }
-
-    get tempP(): number { return this._tempP; }
-
-    get tempE(): number { return this._tempE; }
-
-    get tempAll(): number { return this.temp + this.tempP + this.tempE; }
-
-    static get zero(): ArmorValues { return new ArmorValues({}); }
-}
+import { sizeCombatMod } from "../Size";
+import { ArmorValues } from "./ArmorValues";
+import { FullArmorValuesBuilder, FullFromBaseBuilder } from "./builder";
 
 class FullArmorValues extends ArmorValues {
     private _str: number;
@@ -122,14 +25,14 @@ class FullArmorValues extends ArmorValues {
         bab = 0,
         size = 0,
     }: FullArmorValuesBuilder) {
-        super({armor, shield, dodge, nat, defl, misc, miscP, miscE, temp, tempP, tempE});
+        super({ armor, shield, dodge, nat, defl, misc, miscP, miscE, temp, tempP, tempE });
         this._str = str;
         this._dex = dex;
         this._bab = bab;
         this._size = size;
     }
 
-    static fromBaseValues({base, stats, bab, size}: FullFromBaseBuilder): FullArmorValues {
+    static fromBaseValues({ base, stats, bab, size }: FullFromBaseBuilder): FullArmorValues {
         const {
             armor,
             shield,
@@ -174,7 +77,7 @@ class FullArmorValues extends ArmorValues {
     get touch(): FullArmorValues {
         return new FullArmorValues({
             armor: 0,
-            shield: 0, 
+            shield: 0,
             dodge: this.dodge,
             nat: 0,
             defl: this.defl,
@@ -192,7 +95,7 @@ class FullArmorValues extends ArmorValues {
     get flatFooted(): FullArmorValues {
         return new FullArmorValues({
             armor: this.armor,
-            shield: this.shield, 
+            shield: this.shield,
             dodge: 0,
             nat: this.nat,
             defl: this.defl, // TODO Verify!
@@ -227,7 +130,4 @@ class FullArmorValues extends ArmorValues {
     }
 }
 
-export {
-    ArmorValues,
-    FullArmorValues,
-};
+export { FullArmorValues };
