+++ src/types/Character/ArmorValues/index.ts
@@ -1,0 +1,2 @@
+export { ArmorValues } from './ArmorValues';
+export { FullArmorValues } from './FullArmorValues';
